Patch

+++ src/ua/spalah/bank/models/accounts/SavingAccount.java
@@ -52,7 +52,7 @@ public class SavingAccount implements Account {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 }
